Refactor io.deephaven.processor.functions to io.deephaven.processor.function
